cfg-lexer: save the original content of CFGI_BUFFER inclusions
The lexer mutates the buffer it gets, thus by the time we want to use
it to display errors, it may have been littered by embedded NULs, as
the reason for the mutation is to zero-terminate yytext. Although
this doubles the memory needs for a CfgIncludeLevel, the size of the
buffer should be reasonably small (few kB) and they get destroyed once the
include file is processed.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>